fix: change login-password usage to one token as bearer

+++ src/utils/git_utils.py
@@ -5,17 +5,17 @@ from pathlib import Path
 
 from loguru import logger
 
-# TODO: add credentials request to init script
-GIT_LOGIN = os.getenv('GIT_LOGIN', '')
-GIT_PASS = os.getenv('GIT_PASS', '')
-
 
 async def ensure_repo(repo: str, branch: str, target_dir: Path):
+    git_token = os.getenv('GIT_TOKEN', '')
+    if not git_token:
+        raise RuntimeError('GIT_TOKEN env var is not set')
+
     if target_dir.exists() and check_repo_dir_config(target_dir, repo):
-        await update_repo(repo, branch, target_dir)
+        await update_repo(repo, branch, git_token, target_dir)
     else:
         shutil.rmtree(target_dir, ignore_errors=True)
-        await clone_repo(repo, branch, target_dir)
+        await clone_repo(repo, branch, git_token, target_dir)
 
 def check_repo_dir_config(repo_dir: Path, repo: str):
     git_dir = repo_dir / '.git'
@@ -28,23 +28,22 @@ def check_repo_dir_config(repo_dir: Path, repo: str):
                     return True
     return False
 
-async def clone_repo(repo: str, branch: str, target_dir: Path):
+async def clone_repo(repo: str, branch: str, git_token: str, target_dir: Path):
     logger.info(f'⬇️ Cloning {repo}, branch: {branch} to {target_dir}...')
     command = [
         'git', 'clone',
+        '-c', f'http.extraHeader=Authorization: Bearer {git_token}',
         '--depth', '1',
         '--branch', branch,
         repo, str(target_dir)
     ]
     process = await asyncio.create_subprocess_exec(
-        *command,
-        stdin=asyncio.subprocess.PIPE,
-        stdout=None,
-        stderr=asyncio.subprocess.PIPE,
-    )
-    await asyncio.sleep(1)  # wait for auth query
-    _, stderr = await process.communicate(
-        input=f'{GIT_LOGIN}\n{GIT_PASS}\n'.encode())
+        *command, stderr=asyncio.subprocess.PIPE)
+    _, stderr = await process.communicate()
+
+    # remove token from config
+    unset_token_command = ['git', 'config', '--unset', 'http.extraHeader']
+    await asyncio.create_subprocess_exec(*unset_token_command, cwd=target_dir)
 
     if process.returncode == 0:
         logger.success(f'✅️ Git clone success from {repo}')
@@ -54,7 +53,7 @@ async def clone_repo(repo: str, branch: str, target_dir: Path):
         raise RuntimeError(f'{error}')
 
 
-async def update_repo(repo: str, branch: str, target_dir: Path):
+async def update_repo(repo: str, branch: str, git_token:str, target_dir: Path):
     logger.info(f'⬇️ Updating {repo}, branch: {branch} in {target_dir}...')
     init_commands = [
         ['git', 'reset', '--hard', 'HEAD'],
@@ -63,28 +62,26 @@ async def update_repo(repo: str, branch: str, target_dir: Path):
 
     for cmd in init_commands:
         process = await asyncio.create_subprocess_exec(
-            *cmd,
-            cwd=target_dir,
-            stdout=None,
-            stderr=asyncio.subprocess.PIPE
-        )
+            *cmd, cwd=target_dir, stderr=asyncio.subprocess.PIPE)
         _, stderr = await process.communicate()
+
         if process.returncode != 0:
             logger.error(f'‼️ Git update cmd failed: {cmd}')
             error = stderr.decode()
             raise RuntimeError(f'{error}')
 
-    pull_command = ['git', 'pull', 'origin', branch]
+    pull_command = [
+        'git',
+        '-c', f'http.extraHeader=Authorization: Bearer {git_token}',
+        'pull', 'origin', branch
+    ]
     update_process = await asyncio.create_subprocess_exec(
-        *pull_command,
-        cwd=target_dir,
-        stdin=asyncio.subprocess.PIPE,
-        stdout=None,
-        stderr=asyncio.subprocess.PIPE
-    )
-    await asyncio.sleep(1)  # wait for auth query
-    _, stderr = await update_process.communicate(
-        input=f'{GIT_LOGIN}\n{GIT_PASS}\n'.encode())
+        *pull_command, cwd=target_dir, stderr=asyncio.subprocess.PIPE)
+    _, stderr = await update_process.communicate()
+
+    # remove token from config
+    unset_token_command = ['git', 'config', '--unset', 'http.extraHeader']
+    await asyncio.create_subprocess_exec(*unset_token_command, cwd=target_dir)
 
     if update_process.returncode == 0:
         logger.success(f'✅️ Git repo update success from {repo}')
